move caliper config and processor from track app to caliper_track app + minor fixes
