+++ common/djangoapps/track/shim.py
@@ -2,10 +2,7 @@
 
 
 import json
-import logging
 
-from openedx.features.caliper_tracking.base_transformer import CaliperBaseTransformer
-from track.caliper_config import EVENT_MAPPING
 from .transformers import EventTransformerRegistry
 
 CONTEXT_FIELDS_TO_INCLUDE = [
@@ -17,8 +14,6 @@ CONTEXT_FIELDS_TO_INCLUDE = [
     'referer',
     'accept_language'
 ]
-
-logger = logging.getLogger(__name__)
 
 
 class LegacyFieldMappingProcessor(object):
@@ -113,16 +108,3 @@ class PrefixedEventProcessor(object):
             return
         event.transform()
         return event
-
-
-class CaliperProcessor(object):
-    def __call__(self, event):
-        try:
-            caliper_event = CaliperBaseTransformer(event).transform_event()
-            related_function = EVENT_MAPPING[event.get('event_type')]
-            return related_function(event, caliper_event)
-        except KeyError as ex:
-            logger.exception("Missing transformer method implementation for %s" % event.get('event_type'))
-        except Exception as ex:
-            logger.exception(ex.args)
-            pass
